# docs/swagger.yaml
definitions:
  dto.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  dto.AdminCreateShortLinksResponse:
    properties:
      created:
        type: integer
      job:
        $ref: '#/definitions/dto.AdminShortLinkUploadJobDTO'
      message:
        type: string
      scenario_id:
        type: integer
      skipped:
        type: integer
      total_rows:
        type: integer
    type: object
...
  dto.AdminDownloadShortLinksByScenarioNameRegexRequest:
    properties:
      scenario_name_regex:
        type: string
    required:
    - scenario_name_regex
    type: object
  dto.AdminDownloadShortLinksRangeRequest:
    properties:
      scenario_from:
        type: integer
      scenario_to:
        type: integer
    required:
    - scenario_from
    - scenario_to
    type: object
  dto.AdminDownloadShortLinksRequest:
    properties:
      scenario_id:
        type: integer
    required:
    - scenario_id
    type: object
...
  dto.AdminShortLinkUploadJobDTO:
    properties:
      attempts:
        type: integer
      created:
        type: integer
      id:
        type: string
      last_error:
        type: string
      next_attempt_at:
        type: string
      published:
        type: integer
      scenario_id:
        type: integer
      skipped:
        type: integer
      status:
        type: string
      total_rows:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/admin/short-links/download:
    post:
      consumes:
      - application/json
      parameters:
      - description: Scenario ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AdminDownloadShortLinksRequest'
      produces:
      - text/csv
      responses:
        "200":
          description: CSV file
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Admin Download Short Links by Scenario
      tags:
      - Admin ShortLinks
  /api/v1/admin/short-links/download-with-clicks:
    post:
      consumes:
      - application/json
      parameters:
      - description: Scenario ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AdminDownloadShortLinksRequest'
      produces:
      - text/csv
      responses:
        "200":
          description: CSV file
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Admin Download Short Links With Clicks by Scenario
      tags:
      - Admin ShortLinks
  /api/v1/admin/short-links/download-with-clicks-by-scenario-name:
    post:
      consumes:
      - application/json
      parameters:
      - description: Scenario name regex
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AdminDownloadShortLinksByScenarioNameRegexRequest'
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: Excel file
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Admin Download Short Links With Clicks by Scenario Name Regex (Excel)
      tags:
      - Admin ShortLinks
  /api/v1/admin/short-links/download-with-clicks-range:
    post:
      consumes:
      - application/json
      parameters:
      - description: Scenario range [from, to)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AdminDownloadShortLinksRangeRequest'
      produces:
      - text/csv
      responses:
        "200":
          description: CSV file
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Admin Download Short Links With Clicks by Scenario Range
      tags:
      - Admin ShortLinks
  /api/v1/admin/short-links/upload-csv:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: CSV file with long_link column
        in: formData
        name: file
        required: true
        type: file
      - description: Domain for short links (e.g., https://j0in.ir)
        in: formData
        name: short_link_domain
        required: true
        type: string
      - description: Scenario name for grouping (stored in short_links.scenario_name)
        in: formData
        name: scenario_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.AdminCreateShortLinksResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Admin Upload Short Links CSV
      tags:
      - Admin ShortLinks
...
securityDefinitions:
  Bearer:
    description: 'JWT Bearer token: Bearer <access_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"